enable drag on project items
